moving test autograd to internals and deleted old utils file
